# contracts/openapi.yaml
...
openapi: 3.1.0
info:
  title: VITalWatch CTMS API
  description: '

    Real-time CTMS and pharmacovigilance dashboard for the All India Institute of
    Ayurveda.


    **Demo system. Synthetic data only. No real patient data at any stage.**


    Which endpoint backs which screen (Ishan):


    | Screen | Endpoints |

    |---|---|

    | `/login` | `POST /auth/login`, `GET /auth/me` |

    | `/portfolio` | `GET /api/kpi/portfolio`, `GET /api/studies`, `GET /api/alerts`
    |

    | `/study/[id]` | `GET /api/studies/{id}`, `GET /api/kpi/study/{id}`, `GET /api/enrolment/{id}`,
    `GET /api/sites?study_id=` |

    | `/ae` | `POST /api/ae`, `GET /api/ae`, `POST /api/coding/suggest`, `GET /api/signals`
    |

    | `/audit` | `GET /api/audit`, `GET /api/audit/verify` |

    | `/alerts` | `GET /api/alerts`, `POST /api/alerts/{id}/ack` |

    '
  version: 0.1.0
paths:
...
  /api/ae:
    get:
      tags:
      - pharmacovigilance
      summary: List adverse events
      operationId: list_aes_api_ae_get
      parameters:
      - name: study_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Study Id
      - name: serious
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Serious
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdverseEvent'
                title: Response List Aes Api Ae Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - pharmacovigilance
      summary: Report an AE/SAE
      description: 'PLACEHOLDER: echoes the payload back.


        Sreeja: code the narrative via services.pv.coding, compute the NDCT-2019 deadlines

        via services.pv.timelines from the SERVER clock, then let Caleb''s middleware
        write

        the audit event.'
      operationId: report_ae_api_ae_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdverseEvent'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdverseEvent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/ae/{ae_id}:
    get:
      tags:
      - pharmacovigilance
      summary: One adverse event
      operationId: get_ae_api_ae__ae_id__get
      parameters:
      - name: ae_id
        in: path
        required: true
        schema:
          type: string
          title: Ae Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdverseEvent'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/coding/suggest:
    post:
      tags:
      - pharmacovigilance
      summary: Suggest coded terms
      description: 'PLACEHOLDER: one fixed candidate. Sreeja: top-3 from the curated
        subset.'
      operationId: suggest_coding_api_coding_suggest_post
      parameters:
      - name: narrative
        in: query
        required: true
        schema:
          type: string
          title: Narrative
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodingSuggestion'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/fhir/ResearchStudy/{study_id}:
    get:
      tags:
      - fhir
      summary: FHIR R4 ResearchStudy
      operationId: research_study_api_fhir_ResearchStudy__study_id__get
      parameters:
      - name: study_id
        in: path
        required: true
        schema:
          type: string
          title: Study Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Research Study Api Fhir Researchstudy  Study Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/fhir/AdverseEvent/{ae_id}:
    get:
      tags:
      - fhir
      summary: FHIR R4 AdverseEvent
      operationId: adverse_event_api_fhir_AdverseEvent__ae_id__get
      parameters:
      - name: ae_id
        in: path
        required: true
        schema:
          type: string
          title: Ae Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Adverse Event Api Fhir Adverseevent  Ae Id  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AECausality:
      type: string
      enum:
      - unrelated
      - unlikely
      - possible
      - probable
      - certain
      title: AECausality
      description: WHO-UMC causality categories.
    AEOutcome:
      type: string
      enum:
      - recovered
      - recovering
      - not_recovered
      - recovered_with_sequelae
      - fatal
      - unknown
      title: AEOutcome
    AESeverity:
      type: string
      enum:
      - mild
      - moderate
      - severe
      title: AESeverity
    AdverseEvent:
      properties:
        id:
          type: string
          title: Id
        study_id:
          type: string
          title: Study Id
        site_id:
          type: string
          title: Site Id
        subject_code:
          type: string
          title: Subject Code
        narrative:
          type: string
          title: Narrative
        onset_date:
          type: string
          format: date
          title: Onset Date
        serious:
          type: boolean
          title: Serious
          default: false
        severity:
          $ref: '#/components/schemas/AESeverity'
        causality:
          $ref: '#/components/schemas/AECausality'
        outcome:
          $ref: '#/components/schemas/AEOutcome'
        coded_term:
          anyOf:
          - type: string
          - type: 'null'
          title: Coded Term
        coded_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Coded Code
        coding_confidence:
          anyOf:
          - type: number
          - type: 'null'
          title: Coding Confidence
        coding_source:
          $ref: '#/components/schemas/CodingSource'
          default: uncoded
        suspect_drug:
          anyOf:
          - type: string
          - type: 'null'
          title: Suspect Drug
        drug_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Drug Code
        drug_coding_source:
          $ref: '#/components/schemas/CodingSource'
          default: uncoded
        reported_at:
          type: string
          format: date-time
          title: Reported At
        deadline_24h:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deadline 24H
        deadline_14d:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deadline 14D
        timeline_status:
          $ref: '#/components/schemas/TimelineStatus'
          default: not_applicable
      additionalProperties: false
      type: object
      required:
      - id
      - study_id
      - site_id
      - subject_code
      - narrative
      - onset_date
      - severity
      - causality
      - outcome
      - reported_at
      title: AdverseEvent
...
    CodingCandidate:
      properties:
        term:
          type: string
          title: Term
        code:
          type: string
          title: Code
        score:
          type: number
          title: Score
        source:
          type: string
          title: Source
      type: object
      required:
      - term
      - code
      - score
      - source
      title: CodingCandidate
    CodingSource:
      type: string
      enum:
      - mock
      - semantic
      - meddra
      - uncoded
      title: CodingSource
      description: Where a coded term came from. Shown in the UI — we never imply
        a licensed dictionary.
    CodingSuggestion:
      properties:
        candidates:
          items:
            $ref: '#/components/schemas/CodingCandidate'
          type: array
          title: Candidates
      type: object
      required:
      - candidates
      title: CodingSuggestion
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TimelineStatus:
      type: string
      enum:
      - on_track
      - due_soon
      - breached
      - not_applicable
      title: TimelineStatus
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError